Legacy methods moved out to a legacy class
