+++ sdk/core/src/shared/getOptionsFromEnvironment.ts
@@ -145,7 +145,6 @@ const splitOptions = <TOptions extends SdkOptions>(options: TOptions) => {
     emailIssuerBasePath,
     revocationUrl,
     storageRegion,
-    cognitoUserTokens,
     ...otherOptions
   } = options
 
@@ -167,13 +166,12 @@ const splitOptions = <TOptions extends SdkOptions>(options: TOptions) => {
       revocationUrl,
     },
     storageRegion,
-    cognitoUserTokens,
     otherOptions,
   }
 }
 
 export const getOptionsFromEnvironment = (options: SdkOptions) => {
-  const { accessApiKeyOptions, environmentOptions, storageRegion, cognitoUserTokens, otherOptions } = splitOptions(
+  const { accessApiKeyOptions, environmentOptions, storageRegion, otherOptions } = splitOptions(
     options,
   )
 
@@ -181,7 +179,6 @@ export const getOptionsFromEnvironment = (options: SdkOptions) => {
     basicOptions: getBasicOptionsFromEnvironment(environmentOptions),
     accessApiKey: getAccessApiKeyFromOptions(accessApiKeyOptions),
     storageRegion,
-    cognitoUserTokens,
     otherOptions,
   }
 }
